backend/requirements.txt
# FastAPI
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Validation
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# Security
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
fastapi-csrf-protect==0.3.4

# Redis
redis[hiredis]==5.2.1

# HTTP Client
httpx==0.28.1

# Environment
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0

# Logging
structlog==24.4.0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.8.5
mypy==1.14.1
